src/frontend/requirements.txt
streamlit
requests
pandas
Pillow
matplotlib
plotly
numpy
